+++ src/main/kotlin/com/group4/ticketingservice/controller/UserController.kt
@@ -43,7 +43,7 @@ class UserController(
      * @author MinJun Kim
      */
     @GetMapping("/access_token_info")
-    fun getAccessTokenInfo(@AuthenticationPrincipal userId: Long): ResponseEntity<Map<String, Any>> {
+    fun getAccessTokenInfo(@AuthenticationPrincipal userId: Int): ResponseEntity<Map<String, Any>> {
         val jwt = SecurityContextHolder.getContext().authentication.credentials.toString()
         val expiresInMillis = tokenProvider.parseTokenExpirationTime(jwt)
         val map = mapOf(
